+++ src/ValueProvider/ConfigValueProvider.mjs
@@ -1,0 +1,56 @@
+import { FILE_SUFFIX } from "./FileValueProvider.mjs";
+import { JSON_SUFFIX } from "./JsonValueProvider.mjs";
+
+/** @typedef {import("../FluxConfigApi.mjs").FluxConfigApi} FluxConfigApi */
+
+const CONFIG_KEY = "config";
+
+export class ConfigValueProvider {
+    /**
+     * @type {{[key: string]: *} | null}
+     */
+    #config = null;
+
+    /**
+     * @returns {ConfigValueProvider}
+     */
+    static new() {
+        return new this();
+    }
+
+    /**
+     * @private
+     */
+    constructor() {
+
+    }
+
+    /**
+     * @param {string} key
+     * @param {FluxConfigApi} flux_config_api
+     * @returns {Promise<*>}
+     */
+    async getConfig(key, flux_config_api) {
+        if (key === CONFIG_KEY || key.endsWith(FILE_SUFFIX) || key.endsWith(JSON_SUFFIX)) {
+            return null;
+        }
+
+        return (await this.#getConfig(
+            flux_config_api
+        ))[key] ?? null;
+    }
+
+    /**
+     * @param {FluxConfigApi} flux_config_api
+     * @returns {Promise<{[key: string]: *}>}
+     */
+    async #getConfig(flux_config_api) {
+        this.#config ??= await flux_config_api.getConfig(
+            CONFIG_KEY,
+            {},
+            false
+        );
+
+        return this.#config;
+    }
+}