+++ src/ValueProvider/ValueProvider.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("../FluxConfigApi.mjs").FluxConfigApi} FluxConfigApi */
+
+/**
+ * @typedef {{getConfig: (key: string, flux_config_api: FluxConfigApi) => Promise<*>}} ValueProvider
+ */